+++ .storybook/addons.js
@@ -2,3 +2,4 @@ import '@storybook/addon-actions/register';
 import '@storybook/addon-links/register';
 import '@storybook/addon-a11y/register';
 import '@storybook/addon-viewport/register';
+import 'storybook-addon-intl/register';
