+++ src/components/LanguageWrapper/index.js
@@ -1,21 +1,13 @@
 import React, { Component, Fragment } from 'react';
-import Routes from 'react-static-routes';
-import locale from 'browser-locale';
-import Cookies from 'js-cookie';
 import PropTypes from 'prop-types';
 import { withSiteData } from 'react-static';
+import Routes from 'react-static-routes';
+import { IntlProvider } from 'react-intl';
+import locale from 'browser-locale';
+import Cookies from 'js-cookie';
 import { find } from 'lodash';
 
-// react-intl i18n
-import { IntlProvider, addLocaleData } from 'react-intl';
-import en from 'react-intl/locale-data/en';
-import es from 'react-intl/locale-data/es';
-import vi from 'react-intl/locale-data/vi';
-import ar from 'react-intl/locale-data/ar';
-import messages_en from 'js/i18n/locales/en.json';
-import messages_es from 'js/i18n/locales/es.json';
-import messages_vi from 'js/i18n/locales/vi.json';
-import messages_ar from 'js/i18n/locales/ar.json';
+import localeMessages from 'js/i18n/loadLocaleData';
 import {
   SUPPORTED_LANGUAGES,
   SUPPORTED_LANG_CODES,
@@ -23,13 +15,6 @@ import {
   LANG_COOKIE_EXPIRES,
   DEFAULT_LANG,
 } from 'js/i18n/constants';
-addLocaleData([...en, ...es, ...vi, ...ar]);
-const localeMessages = {
-  en: messages_en,
-  es: messages_es,
-  vi: messages_vi,
-  ar: messages_ar,
-};
 
 // page_sections
 import LanguageSelectBar from 'components/PageSections/LanguageSelectBar';
